feat(optimization): first-class to_precision() with overflow validation
Casting to half precision was a place users had to drop to raw torch
(model.to(dtype)) and hope — fp16's 65504 ceiling silently NaNs attention-heavy
models. Add to_precision(model, "bf16"|"fp16", example_input=...): it warns on the
fp16 overflow risk and, given an example input, runs one forward and asserts the
outputs are finite, raising NumericalInstabilityError (recommending bf16) on
overflow. Also fix assert_finite to cast through float32 so it handles bf16/fp16
tensors (numpy has no bfloat16).

+++ mindtrace/models/mindtrace/models/optimization/export/onnx_export.py
@@ -85,7 +85,8 @@ def assert_finite(outputs: Any, *, context: str) -> None:
     else:
         items = [outputs]
     for index, item in enumerate(items):
-        arr = item.detach().cpu().numpy() if hasattr(item, "detach") else np.asarray(item)
+        # .float() first: numpy has no bfloat16, and it normalizes fp16 too.
+        arr = item.detach().float().cpu().numpy() if hasattr(item, "detach") else np.asarray(item)
         if not np.isfinite(np.asarray(arr, dtype=np.float64)).all():
             raise NumericalInstabilityError(
                 f"{context} produced non-finite (NaN/Inf) values in output {index}. This usually means "
